refactor(query): flatten isomorphic/ to query root
Move all cross-platform query files from query/isomorphic/ to query/.
Merge isomorphic barrel into query/index.ts. Desktop exception stays nested.
The "isomorphic" folder was a misnomer—this app has no server—and
the nesting was inverted (90% of code in a subfolder).
