+++ apps/whispering/src/lib/query/analytics.ts
