+++ apps/whispering/src/lib/query/shortcuts.ts
