+++ apps/whispering/src/lib/query/text.ts
